requirements.txt
keras==2.9.0
matplotlib==3.5.2
numpy==1.26.4
Pillow==9.2.0
Pillow==10.3.0
tensorflow==2.9.1
